+++ src/layouts/index.tsx
@@ -53,17 +53,16 @@ function MainLayout({ children }: layoutProps) {
                 {breadCrumb[0]?.icon ? breadCrumb[0]?.icon : <MenuIcon />}
               </button>
               <div className="w-full flex justify-between">
-                <div className=" flex items-center gap-2">
+                <div className=" flex items-center gap-2 whitespace-nowrap">
                   {breadCrumb.map((crumb, index) => {
                     return (
                       <Link
                         to={crumb?.url}
                         key={crumb?.label}
-                        className="flex items-center gap-1"
+                        className="flex items-center gap-1 hover:text-primary/60 transition-all"
                       >
-                        {index !== 0 && crumb?.icon}{" "}
-                        <h6 className=" text-lg w-full">{crumb?.label}</h6>{" "}
-                        {">"}
+                        {index !== 0 && ">"}
+                        <h6 className=" text-lg w-full">{crumb?.label}</h6>
                       </Link>
                     );
                   })}